requirements.txt
openai-whisper~=20240930
torch~=2.7.0
transformers~=4.52.2
yt-dlp~=2025.5.22